// wpref-frontend/src/app/api/generated/model/language-enum.ts
/**
 * WpRef API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * * `en` - English * `fr` - Français * `nl` - Nederlands
 */
export enum LanguageEnumDto {

    En = 'en',

    Fr = 'fr',

    Nl = 'nl'
}