// wpref-frontend/src/app/api/generated/model/me.ts
/**
 * WpRef API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LanguageEnumDto } from './language-enum';


export interface MeDto { 
    readonly id: number;
    /**
     * Required. 150 characters or fewer. Letters, digits and @/./+/-/_ only.
     */
    readonly username: string;
    email?: string;
    first_name?: string;
    last_name?: string;
    language?: LanguageEnumDto;
    /**
     * Designates that this user has all permissions without explicitly assigning them.
     */
    is_superuser?: boolean;
    /**
     * Designates whether the user can log into this admin site.
     */
    is_staff?: boolean;
    current_domain?: number | null;
    readonly current_domain_title: string;
    readonly owned_domain_ids: Array<number>;
    readonly managed_domain_ids: Array<number>;
}